6.2. fix
